feat(ignore): Refactor legrand.ts
Introduced a Legrand specific library
